pyproject.toml
[tool.poetry]
name = "botsmongodb"
version = "0.1.0"
description = ""
authors = ["Bol4onok <95026332+MihailGulkin@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
aiogram = {url = "https://github.com/aiogram/aiogram/archive/refs/heads/dev-3.x.zip"}
motor = "^3.1.2"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
